Type Alias pallet_asset_manager::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn register_foreign_asset( origin: OriginFor<T>, asset: T::ForeignAssetType, metadata: T::AssetRegistrarMetadata, min_amount: T::Balance, is_sufficient: bool ) -> DispatchResult

Register new asset with the asset manager

pub fn set_asset_units_per_second( origin: OriginFor<T>, asset_type: T::ForeignAssetType, units_per_second: u128, num_assets_weight_hint: u32 ) -> DispatchResult

Change the amount of units we are charging per execution second for a given ForeignAssetType

pub fn change_existing_asset_type( origin: OriginFor<T>, asset_id: T::AssetId, new_asset_type: T::ForeignAssetType, num_assets_weight_hint: u32 ) -> DispatchResult

Change the xcm type mapping for a given assetId We also change this if the previous units per second where pointing at the old assetType

pub fn remove_supported_asset( origin: OriginFor<T>, asset_type: T::ForeignAssetType, num_assets_weight_hint: u32 ) -> DispatchResult

pub fn remove_existing_asset_type( origin: OriginFor<T>, asset_id: T::AssetId, num_assets_weight_hint: u32 ) -> DispatchResult

Remove a given assetId -> assetType association

pub fn destroy_foreign_asset( origin: OriginFor<T>, asset_id: T::AssetId, num_assets_weight_hint: u32 ) -> DispatchResult

Destroy a given foreign assetId The weight in this case is the one returned by the trait plus the db writes and reads from removing all the associated data

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

The account ID of AssetManager

impl<T: Config> Pallet<T>

pub fn asset_id_type<KArg>(k: KArg) -> Option<T::ForeignAssetType>where KArg: EncodeLike<T::AssetId>,

An auto-generated getter for AssetIdType.

impl<T: Config> Pallet<T>

pub fn asset_type_id<KArg>(k: KArg) -> Option<T::AssetId>where KArg: EncodeLike<T::ForeignAssetType>,

An auto-generated getter for AssetTypeId.

impl<T: Config> Pallet<T>

pub fn asset_type_units_per_second<KArg>(k: KArg) -> Option<u128>where KArg: EncodeLike<T::ForeignAssetType>,

An auto-generated getter for AssetTypeUnitsPerSecond.

impl<T: Config> Pallet<T>

pub fn supported_fee_payment_assets() -> Vec<T::ForeignAssetType>

An auto-generated getter for SupportedFeePaymentAssets.