Crate pallet_asset_manager

TODO Doc comments for the pallet

Asset Manager Pallet

This pallet allows to register new assets if certain conditions are met The main goal of this pallet is to allow moonbeam to register XCM assets and control the creation of local assets The assumption is we work with AssetTypes, which can then be comperted to AssetIds

This pallet has five storage items: AssetIdType, which holds a mapping from AssetId->AssetType AssetTypeUnitsPerSecond: an AssetType->u128 mapping that holds how much each AssetType should be charged per unit of second, in the case such an Asset is received as a XCM asset. Finally, AssetTypeId holds a mapping from AssetType -> AssetId. LocalAssetCounter which holds the counter of local assets that have been created so far. And LocalAssetDeposit, which holds a mapping between assetId and assetInfo, i.e., the asset creator (from which we take the deposit) and the deposit amount itself.

This pallet has eight extrinsics: register_foreign_asset, which registers a foreign asset in this pallet and creates the asset as dictated by the AssetRegistrar trait. set_asset_units_per_second: which sets the unit per second that should be charged for a particular asset. change_existing_asset_type: which allows to update the correspondence between AssetId and AssetType remove_supported_asset: which removes an asset from the supported assets for fee payment remove_existing_asset_type: which removes a mapping from a foreign asset to an assetId register_local_asset: which creates a local asset with a specific owner destroy_foreign_asset: which destroys a foreign asset and all its associated data destroy_local_asset: which destroys a local asset and all its associated data

Re-exports

pub use pallet::*;

Modules

migrations

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_asset_manager