Crate pallet_asset_manager

TODO Doc comments for the pallet

Asset Manager Pallet

This pallet allows to register new assets if certain conditions are met The main goal of this pallet is to allow moonbeam to register XCM assets The assumption is we work with AssetTypes, which can then be comperted to AssetIds

This pallet has three storage items: AssetIdType, which holds a mapping from AssetId->AssetType AssetTypeUnitsPerSecond: an AssetType->u128 mapping that holds how much each AssetType should be charged per unit of second, in the case such an Asset is received as a XCM asset. Finally, AssetTypeId holds a mapping from AssetType -> AssetId.

This pallet has three extrinsics: register_asset, which registers an Asset in this pallet and creates the asset as dictated by the AssetRegistrar trait. set_asset_units_per_second: which sets the unit per second that should be charged for a particular asset. change_existing_asset_type: which allows to update the correspondence between AssetId and AssetType

Re-exports

pub use pallet::*;

Modules

migrations

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_asset_manager