Crate pallet_asset_manager

Asset Manager Pallet

This pallet allows to register new assets if certain conditions are met The main goal of this pallet is to allow moonbeam to register XCM assets The assumption is we work with AssetTypes, which can then be compared to AssetIds

This pallet has five storage items: AssetIdType, which holds a mapping from AssetId->AssetType AssetTypeUnitsPerSecond: an AssetType->u128 mapping that holds how much each AssetType should be charged per unit of second, in the case such an Asset is received as a XCM asset. Finally, AssetTypeId holds a mapping from AssetType -> AssetId.

This pallet has eight extrinsics: register_foreign_asset, which registers a foreign asset in this pallet and creates the asset as dictated by the AssetRegistrar trait. set_asset_units_per_second: which sets the unit per second that should be charged for a particular asset. change_existing_asset_type: which allows to update the correspondence between AssetId and AssetType remove_supported_asset: which removes an asset from the supported assets for fee payment remove_existing_asset_type: which removes a mapping from a foreign asset to an assetId destroy_foreign_asset: which destroys a foreign asset and all its associated data

Re-exports

pub use crate::weights::WeightInfo;

pub use pallet::*;

Modules

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_asset_manager