Module pallet_asset_manager::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageAssetIdType

_GeneratedPrefixForStorageAssetTypeId

_GeneratedPrefixForStorageAssetTypeUnitsPerSecond

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

An error that can occur while executing the mapping pallet’s logic.

Event

The event emitted by this pallet.

Constants

PALLET_ID

The AssetManagers’s pallet id

Traits

AssetRegistrar

Config

Configuration trait of this pallet.

Type Definitions

AssetIdType

Mapping from an asset id to asset type. This is mostly used when receiving transaction specifying an asset directly, like transferring an asset from this chain to another.

AssetTypeId

Reverse mapping of AssetIdType. Mapping from an asset type to an asset id. This is mostly used when receiving a multilocation XCM message to retrieve the corresponding asset in which tokens should me minted.

AssetTypeUnitsPerSecond

Stores the units per second for local execution for a AssetType. This is used to know how to charge for XCM execution in a particular asset Not all assets might contain units per second, hence the different storage

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.