Module pallet_asset_manager::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

An error that can occur while executing the mapping pallet’s logic.

Event

The Event enum of this pallet

Constants

PALLET_ID

The AssetManagers’s pallet id

Traits

AssetRegistrar

Config

Configuration trait of this pallet.

Type Aliases

AssetIdType

Mapping from an asset id to asset type. This is mostly used when receiving transaction specifying an asset directly, like transferring an asset from this chain to another.

AssetTypeId

Reverse mapping of AssetIdType. Mapping from an asset type to an asset id. This is mostly used when receiving a multilocation XCM message to retrieve the corresponding asset in which tokens should me minted.

AssetTypeUnitsPerSecond

Stores the units per second for local execution for a AssetType. This is used to know how to charge for XCM execution in a particular asset Not all assets might contain units per second, hence the different storage

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

SupportedFeePaymentAssets

Storage type is [StorageValue] with value type Vec < T :: ForeignAssetType >.