Trait pallet_asset_manager::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen;
    type AssetRegistrarMetadata: Member + Parameter + Default;
    type ForeignAssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen;
    type AssetRegistrar: AssetRegistrar<Self>;
    type ForeignAssetModifierOrigin: EnsureOrigin<Self::Origin>;
    type LocalAssetModifierOrigin: EnsureOrigin<Self::Origin>;
    type LocalAssetIdCreator: LocalAssetIdCreator<Self>;
    type AssetDestroyWitness: Member + Parameter + Copy;
    type Currency: ReservableCurrency<Self::AccountId>;
    type LocalAssetDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen

The Asset Id. This will be used to create the asset and to associate it with a assetType

type AssetRegistrarMetadata: Member + Parameter + Default

The Asset Metadata we want to store

type ForeignAssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default

The Foreign Asset Kind.

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen

The units in which we record balances.

type AssetRegistrar: AssetRegistrar<Self>

The asset Registrar. The trait we use to register Assets

type ForeignAssetModifierOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to create and modify asset information for foreign assets

type LocalAssetModifierOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to create and modify asset information for local assets

type LocalAssetIdCreator: LocalAssetIdCreator<Self>

Ways of creating local asset Ids

type AssetDestroyWitness: Member + Parameter + Copy

The asset destroy Witness

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism in which we reserve deposits for local assets.

type LocalAssetDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The basic amount of funds that must be reserved for a local asset.

type WeightInfo: WeightInfo