Trait pallet_asset_manager::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen;
    type AssetRegistrarMetadata: Member + Parameter + Default;
    type AssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen;
    type AssetRegistrar: AssetRegistrar<Self>;
    type AssetModifierOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen

The Asset Id. This will be used to register the asset in Assets

type AssetRegistrarMetadata: Member + Parameter + Default

The Asset Metadata we want to store

type AssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default

The Asset Kind.

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen

The units in which we record balances.

type AssetRegistrar: AssetRegistrar<Self>

The trait we use to register Assets

type AssetModifierOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to create and modify asset information

type WeightInfo: WeightInfo