Trait pallet_asset_manager::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen;
    type AssetRegistrarMetadata: Member + Parameter + Default;
    type ForeignAssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default;
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen;
    type AssetRegistrar: AssetRegistrar<Self>;
    type ForeignAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type LocalAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type LocalAssetIdCreator: LocalAssetIdCreator<Self>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type LocalAssetDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type AssetId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen

The Asset Id. This will be used to create the asset and to associate it with a assetType

type AssetRegistrarMetadata: Member + Parameter + Default

The Asset Metadata we want to store

type ForeignAssetType: Parameter + Member + Ord + PartialOrd + Into<Self::AssetId> + Default

The Foreign Asset Kind.

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen

The units in which we record balances.

type AssetRegistrar: AssetRegistrar<Self>

The asset Registrar. The trait we use to register Assets

type ForeignAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to create and modify asset information for foreign assets

type LocalAssetModifierOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin that is allowed to create and modify asset information for local assets

type LocalAssetIdCreator: LocalAssetIdCreator<Self>

Ways of creating local asset Ids

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism in which we reserve deposits for local assets.

type LocalAssetDeposit: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The basic amount of funds that must be reserved for a local asset.

type WeightInfo: WeightInfo